move network policy related code to another file
